Fix compilation warnings.
git-svn-id: https://sourceforge.llnl.gov/svn/repos/samrai_project/trunk/MASTER@902 f45145c3-882e-4954-aa44-00140844b9ef

+++ source/SAMRAI/mesh/GriddingAlgorithmConnectorWidthRequestor.C
@@ -47,7 +47,7 @@ GriddingAlgorithmConnectorWidthRequestor::computeRequiredConnectorWidths(
    const hier::PatchHierarchy& patch_hierarchy) const
 {
    if ( !d_tag_to_cluster_width.empty() ) {
-      TBOX_ASSERT( d_tag_to_cluster_width.size() >=
+      TBOX_ASSERT( static_cast<int>(d_tag_to_cluster_width.size()) >=
                    patch_hierarchy.getMaxNumberOfLevels()-1 );
    }
